feat: reducing first load js on homepage
