+++ app/page.tsx
@@ -1,356 +1,38 @@
-/* eslint-disable @next/next/no-img-element */
-import React, { Suspense, useState } from "react";
-import Link from "next/link";
-import { PlusIcon } from "@heroicons/react/24/solid";
-import { CloudArrowUpIcon, LockClosedIcon } from "@heroicons/react/20/solid";
-import { CheckIcon } from "@heroicons/react/24/outline";
-import { Button } from "@/components/Utilities/Button";
+// /* eslint-disable @next/next/no-img-element */
+
+import { Suspense } from "react";
+// import { Presentation } from "@/components/Pages/Home/Presentation";
 import dynamic from "next/dynamic";
-import { chosenCommunities } from "@/utilities/chosenCommunities";
-import { PAGES } from "@/utilities/pages";
 import { defaultMetadata } from "@/utilities/meta";
-import { ImageTheme } from "@/components/Utilities/ImageTheme";
-import { TypedLoading } from "@/components/Pages/Home/ReactTypedWrapp";
-// import { ReactTypedWrapper } from "@/components/Pages/Home/ReactTypedWrapp";
-import { MegaphoneIcon } from "@/components/Icons/Megaphone";
-import { RightArrowIcon } from "@/components/Icons/RightArrow";
-import { ExternalLink } from "@/components/Utilities/ExternalLink";
-
-const ProjectDialog = dynamic(
-  () =>
-    import("@/components/Dialogs/ProjectDialog/index").then(
-      (mod) => mod.ProjectDialog
-    ),
-  { ssr: false }
-);
-const ReactTypedWrapper = dynamic(
-  () =>
-    import("@/components/Pages/Home/ReactTypedWrapp").then(
-      (mod) => mod.ReactTypedWrapper
-    ),
-  { ssr: false, loading: () => <TypedLoading /> }
-);
-
-const Presentation = () => {
-  return (
-    <div className="flex flex-1 items-end gap-8 max-2xl:gap-4">
-      <div className=" flex w-full flex-row items-end justify-center gap-1 max-lg:mt-4">
-        <div className="flex w-full flex-[4]  flex-col  items-start">
-          <div className="flex flex-1 flex-col gap-8 pb-8 max-2xl:gap-6">
-            <h4 className="w-max rounded-3xl bg-[#EAECF5] dark:bg-slate-800 dark:text-blue-400 px-3 py-1 text-center text-base font-semibold text-indigo-900  max-2xl:text-sm">
-              ONCHAIN PROTOCOL
-            </h4>
-
-            <div className="flex flex-col gap-1 pb-8 max-2xl:pb-1">
-              <h1 className="text-6xl font-bold leading-[64px] text-gray-900 dark:text-gray-100 max-2xl:text-4xl max-2xl:leading-[36px] max-lg:text-3xl max-lg:leading-normal">
-                Visibility and <br />
-                accountability for <br className="mb-3" />
-                <ReactTypedWrapper />
-                <span className="text-6xl font-bold leading-[64px] text-gray-900 dark:text-gray-100 max-2xl:text-4xl max-2xl:leading-[36px] max-lg:text-3xl max-lg:leading-normal">
-                  <br /> projects
-                </span>
-              </h1>
-            </div>
-
-            <div className="w-max">
-              <ProjectDialog
-                buttonElement={{
-                  icon: (
-                    <img
-                      className="h-6 w-6 text-white"
-                      alt="Contact"
-                      src="/icons/arrow-right-2.svg"
-                    />
-                  ),
-                  iconSide: "right",
-                  text: "Add your project",
-                  styleClass:
-                    "flex rounded-md hover:opacity-75 border-none transition-all ease-in-out duration-300 items-center h-max w-max flex-row gap-3 bg-brand-darkblue dark:bg-gray-700 px-7 py-4 text-lg font-semibold leading-7 text-white hover:bg-brand-darkblue max-2xl:px-5 max-2xl:text-base max-lg:text-sm",
-                }}
-              />
-            </div>
-          </div>
-
-          <div className="flex w-full flex-1 flex-row flex-wrap gap-6 max-md:flex-col">
-            <div className="flex  min-w-min max-w-[500px] flex-1 flex-col items-start gap-3 rounded-3xl  text-gray-900 dark:text-gray-900 bg-[#D7F8EF] px-8 py-6 max-2xl:px-6 max-lg:min-w-full">
-              <div className="flex items-center justify-center rounded-lg border border-black p-2 ">
-                <img
-                  src="/icons/coins-stacked.png"
-                  alt="Grantee"
-                  className="h-7 w-7"
-                />
-              </div>
-              <h2 className="text-2xl text-black font-bold max-2xl:text-xl">
-                For Grantees
-              </h2>
-              <ul className="text-gray-900">
-                <li className="flex w-full flex-row gap-4 break-words">
-                  <CheckIcon className="h-6 min-h-[24px] w-6 min-w-[24px]" />
-                  <p className="text-base font-normal">
-                    Self-Report on grant progress.
-                  </p>
-                </li>
-                <li className="flex w-full flex-row gap-4 break-words">
-                  <CheckIcon className="h-6 min-h-[24px] w-6 min-w-[24px]" />
-                  <p className="text-base font-normal">Build Reputation.</p>
-                </li>
-                <li className="flex w-full flex-row gap-4 break-words">
-                  <CheckIcon className="h-6 min-h-[24px] w-6 min-w-[24px]" />
-                  <p className="text-base font-normal">Receive more funding.</p>
-                </li>
-              </ul>
-            </div>
-            <div className="flex  min-w-min max-w-[500px] flex-1 flex-col items-start gap-3 rounded-3xl text-gray-900 dark:text-gray-900 bg-[#D7F8EF] px-8 py-6 max-2xl:px-6 max-lg:min-w-full">
-              <div className="flex items-center justify-center rounded-lg border border-black p-2">
-                <img
-                  src="/icons/globe.png"
-                  alt="For Community + Grant Admins"
-                  className="h-7 w-7"
-                />
-              </div>
-              <h2 className="w-full text-black break-words text-2xl font-bold  max-2xl:text-xl">
-                For Community + Grant Admins
-              </h2>
-              <ul className="text-gray-900">
-                <li className="flex w-full flex-row gap-4 break-words">
-                  <CheckIcon className="h-6 min-h-[24px] w-6 min-w-[24px]" />
-                  <p className="text-base font-normal ">
-                    Stay up-to-date on projects progress.
-                  </p>
-                </li>
-                <li className="flex w-full flex-row gap-4 break-words">
-                  <CheckIcon className="h-6 min-h-[24px] w-6 min-w-[24px]" />
-                  <p className="text-base font-normal ">
-                    Trigger payments based on milestone completion.
-                  </p>
-                </li>
-                <li className="flex w-full flex-row gap-4 break-words">
-                  <CheckIcon className="h-6 max-h-[24px] min-h-[24px] w-6 min-w-[24px] max-w-[24px]" />
-                  <p className="text-base font-normal ">
-                    Endorse projects, monitor fund usage and flag discrepancies.
-                  </p>
-                </li>
-              </ul>
-            </div>
-            <div />
-          </div>
-        </div>
-        <div className="flex h-full w-full max-w-[720px] flex-col justify-end max-2xl:max-w-[500px] max-xl:max-w-[360px] max-lg:hidden">
-          <img
-            className="aspect-auto h-[auto] w-full max-w-[720px]"
-            src="/images/homepage-artwork.png"
-            alt="Homepage artwork"
-          />
-        </div>
-      </div>
-    </div>
-  );
-};
-
-const Communities = () => {
-  // const { communities } = useCommunity();
-
-  // const communitiesNameToShow = [
-  //   'optimism',
-  //   'gitcoin',
-  //   'arbitrum',
-  //   'ethereum-foundation',
-  // ];
-
-  // const communitiesToShow = communities.filter((community) =>
-  //   communitiesNameToShow.includes(
-  //     community.details?.slug ||
-  //       community.details?.name.toLowerCase() ||
-  //       community.uid
-  //   )
-  // );
-
-  return (
-    <div className="mb-8 mt-[80px] flex h-max w-full flex-row justify-center gap-4 max-md:flex-col max-lg:flex-wrap">
-      <div className="flex h-max w-max flex-col gap-6 rounded-3xl  bg-[#EAECF0] dark:bg-zinc-400 p-8 max-lg:w-full">
-        <div className="h-max w-max rounded-lg border border-black p-2">
-          <img src="/icons/globe.svg" alt="Globe" className="h-7 w-7" />
-        </div>
-        <h4 className="text-5xl font-bold leading-[50px] text-gray-900 max-lg:text-3xl">
-          Communities using GAP
-        </h4>
-      </div>
-      <div className="flex w-full flex-col items-center justify-center gap-8">
-        <div className="grid w-full grid-cols-4 max-md:grid-cols-2">
-          {chosenCommunities().map((community) => (
-            <Link
-              key={community.uid}
-              href={PAGES.COMMUNITY.ALL_GRANTS(community.slug || community.uid)}
-              className="flex hover:shadow-md h-full max-h-[166px] w-full flex-col items-center justify-center gap-2 rounded-2xl px-4 py-6 transition-all duration-300 ease-in-out "
-            >
-              <ImageTheme
-                alt={community.name}
-                lightSrc={community.imageURL.light}
-                darkSrc={community.imageURL.dark}
-                className={
-                  "h-20 w-20 rounded-full object-cover max-lg:h-10 max-lg:w-10 max-sm:h-6 max-sm:w-6"
-                }
-              />
-
-              <p className="line-clamp-1 text-center text-xl font-bold text-black dark:text-zinc-400 max-sm:text-base">
-                {community.name}
-              </p>
-            </Link>
-          ))}
-          <a
-            href="https://tally.so/r/wd0jeq"
-            target="_blank"
-            rel="noreferrer"
-            className="flex h-[166px] w-full flex-col items-center justify-center gap-2 rounded-2xl px-4 py-6 transition-all duration-300 ease-in-out"
-          >
-            <Button className="flex flex-row items-center gap-2 dark:border-slate-600 border border-black bg-white dark:bg-zinc-700 dark:text-white rounded-md p-4 text-sm font-semibold text-black hover:bg-white max-lg:text-xs">
-              Add your Community
-              <PlusIcon className="h-5 w-5" />
-            </Button>
-          </a>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-const WhatIsSolving = () => {
-  return (
-    <div className="mb-10 mt-16 flex flex-row flex-wrap gap-16 max-lg:flex-col">
-      <div className="flex max-w-full flex-1 flex-col gap-2.5 max-2xl:max-w-lg max-lg:max-w-full">
-        <h4 className="w-max rounded-3xl bg-[#EAECF5] dark:bg-slate-800 dark:text-blue-400 px-3 py-1 text-center text-base font-semibold text-indigo-900  max-2xl:text-sm">
-          100% ONCHAIN
-        </h4>
-        <h2 className="text-4xl font-bold text-gray-900 max-sm:text-2xl dark:text-zinc-100">
-          Why are we building this?
-        </h2>
-        <p className="text-xl font-normal text-gray-900 max-sm:text-base dark:text-zinc-200">
-          Annually, the crypto ecosystem issues grants amounting to millions of
-          dollars. While this funding is crucial for ecosystem growth, it has
-          also introduced a range of issues.
-          <br />
-          <br />
-          The Grantee Accountability Protocol (GAP) is designed to address these
-          challenges by aiding grantees in building their reputation, assisting
-          communities in maintaining grantee accountability, and enabling third
-          parties to develop applications using this protocol.
-        </p>
-      </div>
-      <div className="flex flex-1 flex-col gap-6">
-        <div className="flex flex-row flex-wrap gap-6">
-          <div className="flex min-w-[280px]  flex-1 flex-col items-start gap-3 rounded-3xl  bg-[#D7F8EF] px-8 py-6 max-2xl:px-6">
-            <div className="flex items-center justify-center rounded-lg border border-black p-2 ">
-              <img
-                src="/icons/coins-stacked.png"
-                alt="Grantee"
-                className="h-7 w-7"
-              />
-            </div>
-            <h3 className="text-2xl font-bold text-black  max-2xl:text-xl">
-              Limited Accessibility
-            </h3>
-            <p className="text-gray-700">
-              Currently, it is challenging for grant teams and the community to
-              easily access and track project progress and milestones, as
-              information is scattered across forums and external links.
-            </p>
-          </div>
-          <div className="flex min-w-[280px]  flex-1 flex-col items-start gap-3 rounded-3xl bg-[#E0EAFF] px-8 py-6 max-2xl:px-6">
-            <div className="flex items-center justify-center rounded-lg border border-black p-2 ">
-              <img
-                src="/icons/coins-stacked.png"
-                alt="Grantee"
-                className="h-7 w-7"
-              />
-            </div>
-            <h3 className=" text-2xl font-bold text-black  max-2xl:text-xl">
-              Reputation Portability
-            </h3>
-            <p className="text-gray-700">
-              Grantees who apply for grants from multiple organizations struggle
-              to establish and carry their reputation consistently across the
-              ecosystem. This is particularly difficult for individuals who are
-              new to the ecosystem and need opportunities to showcase their work
-              and build their reputation.
-            </p>
-          </div>
-        </div>
-        <div className="flex min-w-[280px] flex-1 flex-col items-start gap-3 rounded-3xl bg-[#EAECF0] px-8 py-6 max-2xl:px-6">
-          <div className="flex items-center justify-center rounded-lg border border-black p-2 ">
-            <img
-              src="/icons/coins-stacked.png"
-              alt="Grantee"
-              className="h-7 w-7"
-            />
-          </div>
-          <h3 className=" text-2xl font-bold text-black  max-2xl:text-xl">
-            Inadequate Data Structure
-          </h3>
-          <p className="text-gray-700">
-            The absence of structured data that can be accessed in a
-            permissionless manner hampers the development of applications and
-            analytical tools for evaluating grant impact and builder reputation.
-          </p>
-        </div>
-      </div>
-    </div>
-  );
-};
+import { Presentation } from "@/components/Pages/Home/Presentation";
+import { NewFeatureBanner } from "@/components/Pages/Home/NewFeatureBanner";
+import { Communities } from "@/components/Pages/Home/Communities";
+import { WhatIsSolving } from "@/components/Pages/Home/WhatIsSolving";
 
 export const metadata = defaultMetadata;
-function NewFeatureBanner() {
-  return (
-    <div className="flex w-full">
-      <div className="flex w-full justify-between bg-[#bee1d8] border-l-[5px] border-[#1de9b6] rounded-l-lg p-4 gap-4 max-md:p-2 max-md:flex-col">
-        <div className="flex flex-row gap-4 items-center max-md:gap-2.5">
-          <MegaphoneIcon />
-          <div className="flex flex-col">
-            <p className="text-sm font-semibold text-[#080a0e] max-md:text-xs">
-              We just launched onchain grant program registry - find funding for
-              your project.
-            </p>
-            {/* <p>
-							Visibility and accountability for community funded and public
-							goods projects.
-						</p> */}
-          </div>
-        </div>
-        <ExternalLink href="/funding-map">
-          <button
-            type="button"
-            className="max-md:text-xs max-md:p-[8px_12px] bg-[#080a0e] rounded-[4px] text-[#1de9b6] flex items-center justify-center gap-[8px] p-[16px_24px] outline-none border-none font-semibold text-[14px] leading-[16px]"
-          >
-            View details
-            <RightArrowIcon />
-          </button>
-        </ExternalLink>
-      </div>
-    </div>
-  );
-}
 
 export default function Index() {
-  const projectTypes = [
-    "Open Source",
-    "Public Goods",
-    "Community Funded",
-    "Crypto protocol",
-  ];
+  // const projectTypes = [
+  //   "Open Source",
+  //   "Public Goods",
+  //   "Community Funded",
+  //   "Crypto protocol",
+  // ];
 
-  const features = [
-    {
-      name: "For Grantees",
-      description:
-        "Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et magna sit morbi lobortis.",
-      icon: CloudArrowUpIcon,
-    },
-    {
-      name: "For Community + Grant Admins",
-      description:
-        "Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et magna sit morbi lobortis.",
-      icon: LockClosedIcon,
-    },
-  ];
+  // const features = [
+  //   {
+  //     name: "For Grantees",
+  //     description:
+  //       "Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et magna sit morbi lobortis.",
+  //     icon: CloudArrowUpIcon,
+  //   },
+  //   {
+  //     name: "For Community + Grant Admins",
+  //     description:
+  //       "Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et magna sit morbi lobortis.",
+  //     icon: LockClosedIcon,
+  //   },
+  // ];
 
   return (
     <>
